requirements.txt
pyqlib
lightgbm
numpy
pandas
matplotlib
jupyter
torch
